navi修正
