+++ app/navigation/__homeTabNav.tsx
@@ -7,7 +7,7 @@ import type { TypeRootList } from '../lib/types';
 import type React from 'react';
 
 /* --------------------------------------------------
- * Home（およびhomeTab）配下の各画面追加
+ * Home配下（およびhomeTab） 各画面追加
  * ----------------------------------------------- */
 
 const HomeTabNav: React.FC = () => {
@@ -16,18 +16,23 @@ const HomeTabNav: React.FC = () => {
 
   return (
     <NestStack.Navigator>
+      {/* --------------------------------------
+       * Home配下（およびhomeTab） 各画面
+       * -------------------------------------- */}
       <NestStack.Screen name='homeTab' options={{ headerShown: false }}>
         {() => (
           <NestTab.Navigator screenOptions={{ swipeEnabled: true }}>
-            {/* --------------------------------------
-             * Home（およびhomeTab）配下の各画面
-             * -------------------------------------- */}
+            {/* homeChild00 画面 */}
             <NestTab.Screen name='homeChild00' options={{ title: 'HomeChild00' }}>
               {(props) => <HomeChild00Screen {...props} />}
             </NestTab.Screen>
+
+            {/* homeChild01 画面 */}
             <NestTab.Screen name='homeChild01' options={{ title: 'HomeChild01' }}>
               {(props) => <HomeChild01Screen {...props} />}
             </NestTab.Screen>
+
+            {/* homeChild02 画面 */}
             <NestTab.Screen name='homeChild02' options={{ title: 'HomeChild02' }}>
               {(props) => <HomeChild02Screen {...props} />}
             </NestTab.Screen>
